requirements.txt
asgiref==3.6.0
attrs==22.2.0
colorama==0.4.6
Django==3.2
django-environ==0.10.0
django-extensions==3.2.1
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==18.2.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
pytest==7.2.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
